ml_service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
xgboost
psycopg2-binary
python-dotenv
